Added a ton of structure.
## LOGIN
Is not working with basics. I still have to do the register, but all of the login and logout function is there. All commented and looking fine!

## GENERAL
All of the basics for the stories, profile views, account page and such are all set up. The infrastructure of the backend code is all set up with comments and looking good. We have the MY_Controller initialized as well for useful and reusable universal functions